+++ clinicadmin/src/main/java/com/clinicadmin/dto/TheraphyProgramWithTheraphyNamesDto.java
@@ -12,7 +12,7 @@ public class TheraphyProgramWithTheraphyNamesDto {
 	
 	private String id;
 	private String programName;
-	private double totalProgramAmount;
+	private int totalProgramAmount;
 	private List<TheraphyNamesDTO> therophy;
 	private String clinicId;
 	private String branchId;
